+++ CHANGELOG.md
@@ -36,6 +36,7 @@ The format is based on [Keep a Changelog], and this project adheres to
   onto multiple tiles if necessary. (\#331)
 * In the new ``mapping`` field of ``RPUConfig`` the max tile input and
   output sizes can be configured for the ``*Mapped`` layers. (\#331)
+* Analog information summary function. (\#316)
 
 ### Fixed